requirements.txt
dash
pandas
PyGithub